+++ src/views/UserEdit/UserEdit.js
@@ -60,11 +60,30 @@ class UserEdit extends React.Component {
     setTenantId: PropTypes.func,
     isCreateKeycloakUserConfirmationOpen: PropTypes.bool,
     initialAssignedRoleIds: PropTypes.object,
+    isLoadingAffiliationRoles: PropTypes.bool.isRequired,
     checkAndHandleKeycloakAuthUser: PropTypes.func,
     confirmCreateKeycloakUser: PropTypes.func
   }
 
   static contextType = CalloutContext;
+
+  state = {
+    isLoading: true,
+  };
+
+  componentDidUpdate() {
+    const { resources } = this.props;
+    const { isLoading } = this.state;
+
+    const isDataLoading = !resourcesLoaded(resources, ['uniquenessValidator']) || (!this.getUser() && this.props.match.params.id);
+
+    // Show loading only when opening the page, because some resources are re-fetched during form save,
+    // which remounts custom fields components, triggering form.change(), which makes the form dirty
+    // and prevents redirection by showing the "Are you sure?" modal.
+    if (isLoading && !isDataLoading) {
+      this.setState({ isLoading: false });
+    }
+  }
 
   getUser() {
     const { resources, match: { params: { id } } } = this.props;
@@ -280,7 +299,7 @@ class UserEdit extends React.Component {
     });
   }
 
-  update({ requestPreferences, readingRoomsAccessList, ...userFormData }) {
+  async update({ requestPreferences, readingRoomsAccessList, ...userFormData }) {
     const {
       updateProxies,
       updateSponsors,
@@ -346,7 +365,7 @@ class UserEdit extends React.Component {
       data.active = (dayjs(user.expirationDate).endOf('day').isSameOrAfter(today));
     }
 
-    this.updateUserData(data, user);
+    await this.updateUserData(data);
   }
 
   async updateUserData(data) {
@@ -483,12 +502,14 @@ class UserEdit extends React.Component {
       setTenantId,
       tenantId,
       setAssignedRoleIds,
+      isLoadingAffiliationRoles,
       assignedRoleIds
     } = this.props;
+    const { isLoading } = this.state;
 
     const profilePictureConfig = get(resources, 'settings.records[0]');
 
-    if (!resourcesLoaded(resources, ['uniquenessValidator']) || (!this.getUser() && this.props.match.params.id)) {
+    if (isLoading) {
       return (
         <LoadingView
           data-test-form-page
@@ -525,6 +546,7 @@ class UserEdit extends React.Component {
         tenantId={tenantId}
         setAssignedRoleIds={setAssignedRoleIds}
         assignedRoleIds={assignedRoleIds}
+        isLoadingAffiliationRoles={isLoadingAffiliationRoles}
       />
     );
   }