+++ test/jest/__mock__/stripesComponents.mock.js
@@ -250,4 +250,5 @@ jest.mock('@folio/stripes/components', () => ({
     );
   }),
   exportToCsv: jest.fn(),
+  LoadingView: jest.fn(() => <div>LoadingView</div>),
 }));